Convert core UOM types from f32 to u32 - fixes all command and data object modules

+++ src/commands/requested/current_requested.rs
@@ -48,7 +48,7 @@ impl CurrentRequested {
             .raw_current()
             .checked_mul(Self::CURRENT_RESOLUTION)
             .ok_or(Ap33772sError::ConversionFailed)?;
-        Ok(ElectricCurrent::new::<milliampere>(f32::from(
+        Ok(ElectricCurrent::new::<milliampere>(u32::from(
             scaled_current,
         )))
     }
