+++ src/commands/requested/voltage_requested.rs
@@ -47,7 +47,7 @@ impl VoltageRequested {
             .raw_voltage()
             .checked_mul(Self::VOLTAGE_RESOLUTION)
             .ok_or(Ap33772sError::ConversionFailed)?;
-        Ok(ElectricPotential::new::<millivolt>(f32::from(
+        Ok(ElectricPotential::new::<millivolt>(u32::from(
             scaled_voltage,
         )))
     }
